cgo is overkill
