--- memprot/memprot_windows.go
@@ -1,3 +1,0 @@
-// +build windows
-
-package memprot
